pyproject.toml
[project]
name = "spectacular-fsize-bug"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Jennings Zhang", email = "dev@sl.jennin.xyz" }
]
dependencies = [
    "django>=5.1.1",
    "drf-spectacular>=0.27.2",
    "djangorestframework>=3.15.2",
    "markdown>=3.7",
    "django-filter>=24.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/spectacularfsizebug"]
